Move video functionality to essential_abilities.py
All video-related methods have been transferred from video.py to essential_abilities.py for better organization and consolidation. The video.py extension now only references that functionality has moved. Required dependencies 'python-pptx' and 'moviepy' were added to docker-requirements.txt.
