+++ docker-requirements.txt
@@ -1,4 +1,5 @@
 pyodbc
+python-pptx
 pyvirtualdisplay
 requests-toolbelt
 solana
@@ -10,6 +11,7 @@ blinkpy>=0.23.0
 ring_doorbell>=0.9.13
 aiohttp
 hikvisionapi>=0.3.2
+moviepy
 azure-ai-inference
 djitellopy
 axis
